requirements.txt
Flask
numpy==1.16.4
tensorflow
keras
redis
rq
pillow
firebase_admin
gunicorn